refactored out sources and models into their own modules It was getting unweildy.
change models_in_dag_order to nodes_in_dag_order to allow for other node types
snapshot/seed
